move shared assets to top level
