+++ src/BookNavigator/BookNavigator.js
@@ -7,8 +7,7 @@ import SharingProvider from '../ItemNavigator/providers/sharing.js';
 import BookmarksProvider from './providers/bookmarks.js';
 import BRFullscreenMgr from './br-fullscreen-mgr.js';
 import { Book } from './BookModel.js';
-// import '../ItemNavigator/ItemNavigator.js';
-import bookLoader from './book-loader.js';
+import bookLoader from './styles/book-loader.js';
 
 const events = {
   menuUpdated: 'menuUpdated',
