+++ src/BookNavigator/styles/bookmark-colors.js
