+++ src/BookNavigator/styles/icon_checkmark.js
