+++ src/BookNavigator/styles/icon_close.js
