+++ src/iaux-book-visual-adjustments/ia-book-visual-adjustments.js
@@ -1,7 +1,7 @@
 import { css, html, LitElement } from "lit-element";
 import { repeat } from "lit-html/directives/repeat.js";
 import { nothing } from "lit-html";
-import checkmarkIcon from './styles/icon_checkmark.js';
+import checkmarkIcon from '../BookNavigator/styles/icon_checkmark.js';
 import "@internetarchive/icon-magnify-minus/icon-magnify-minus";
 import "@internetarchive/icon-magnify-plus/icon-magnify-plus";
